Refs #11733 Fix compiler warnings
