+++ Code/Mantid/Framework/Algorithms/inc/MantidAlgorithms/ChangeTimeZero.h
@@ -70,21 +70,6 @@ private:
   const std::string m_defaultAbsoluteTimeShift;
 };
 
-
-/**
- * General check if we are dealing with a time series
- * @param prop :: the property which is being checked
- * @return True if the proerpty is a time series, otherwise false.
- */
-bool isTimeSeries(Mantid::Kernel::Property *prop) {
-  auto isTimeSeries = false;
-  if (dynamic_cast<Mantid::Kernel::ITimeSeriesProperty *>(prop)) {
-    isTimeSeries = true;
-  }
-  return isTimeSeries;
-}
-
-
 } // namespace Mantid
 } // namespace Algorithms
